Update app.py
chore(app.py): Update float type of Lat and Long app.py

+++ app.py
@@ -178,8 +178,8 @@ data=data.rename(columns={'Lat.':'Lat','Long.':'Long',
                           'Historical Period':'Period',
                           'Date mean in BP in years before 1950 CE [OxCal mu for a direct radiocarbon date, and average of range for a contextual date]':'Date'})
 
-#data["Lat"] = data["Lat"].str.replace(",", ".").astype(float) 
-#data["Long"] = data["Long"].str.replace(",", ".").astype(float) 
+data["Lat"] = data["Lat"].astype(float) 
+data["Long"] = data["Long"].astype(float) 
 #data["PRS_SCZ"] = data["PRS_SCZ"].str.replace(",", ".").astype(float) 
 #data=data.drop(data[data.Region=="Indeterminado"].index) 
 data["mtdna"] = data["mtdna"].str.replace('n/a (<2x)', "..")
